requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.4
pypika-tortoise==0.1.6
pytz==2022.7.1
tortoise-orm==0.19.3
typing_extensions==4.4.0
yarl==1.8.2
